Cleanup BufWriter
